move odoo snippet commands to dom helpers
