+++ packages/OdooWebsiteEditor/odoo-integration.ts
@@ -1,0 +1,22 @@
+import { BasicEditor } from '../bundle-basic-editor/BasicEditor';
+import { DevTools } from '../plugin-devtools/src/DevTools';
+import { OdooWebsiteEditor } from '../bundle-odoo-website-editor/OdooWebsiteEditor';
+import { VRange, withRange } from '../core/src/VRange';
+import { DomLayoutEngine } from '../plugin-dom-layout/src/ui/DomLayoutEngine';
+import { Layout } from '../plugin-layout/src/Layout';
+import { Renderer } from '../plugin-renderer/src/Renderer';
+import { ImageNode } from '../plugin-image/src/ImageNode';
+import { getOdooCommands } from '../plugin-odoo/src/OdooCommands';
+
+export {
+    OdooWebsiteEditor,
+    BasicEditor,
+    DevTools,
+    Layout,
+    DomLayoutEngine,
+    Renderer,
+    ImageNode,
+    withRange,
+    VRange,
+    getOdooCommands,
+};
